requirements.txt
dash
dash-bootstrap-components
pandas
plotly
databricks-sql-connector
databricks-sdk
python-dotenv
dash-ag-grid
streamlit
langchain-community>=0.2.12
pyarrow
